Navigation: Refactor and clean up code
