+++ lib/screens/profile_setup/applicant/disabilities_screen.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:pronto/widgets/progress_indicator.dart';
 import 'package:pronto/constants.dart';
-import 'location_screen.dart';
+import 'package:pronto/router.dart';
 
 class DisabilityScreen extends StatefulWidget {
   final String? userId;
@@ -46,12 +46,7 @@ class _DisabilityScreenState extends State<DisabilityScreen> {
 
       if (!mounted) return;
 
-      Navigator.push(
-        context,
-        MaterialPageRoute(
-          builder: (context) => LocationScreen(userId: widget.userId),
-        ),
-      );
+      NavigationHelper.navigateTo('/location', arguments: widget.userId);
     } catch (e) {
       ScaffoldMessenger.of(
         context,
